+++ source/redux/HttpClient.js
@@ -1,13 +1,13 @@
 import HttpClient from '../HttpClient'
 
 export default function createHttpClient(settings, getStore, options = {}) {
-	let on_before_send
-	let catch_to_retry
-	let get_access_token
+	let onBeforeSend
+	let catchToRetry
+	let getAuthToken
 
 	// Add `getState()` to `http.onRequest()` parameters.
 	if (settings.http.onRequest) {
-		on_before_send = (request, parameters) => {
+		onBeforeSend = (request, parameters) => {
 			settings.http.onRequest(request, {
 				...parameters,
 				getState: getStore().getState
@@ -17,7 +17,7 @@ export default function createHttpClient(settings, getStore, options = {}) {
 
 	// Add `store` and `http` helpers to `http.catch`
 	if (settings.http.catch) {
-		catch_to_retry = (error, retryCount, helpers) => {
+		catchToRetry = (error, retryCount, helpers) => {
 			return settings.http.catch(error, retryCount, helpers)
 			// {
 			// 	...helpers,
@@ -28,7 +28,7 @@ export default function createHttpClient(settings, getStore, options = {}) {
 
 	// Add `store` helper to `authentication.accessToken`
 	if (settings.authentication.accessToken) {
-		get_access_token = (getCookie, helpers) => {
+		getAuthToken = (getCookie, helpers) => {
 			return settings.authentication.accessToken({
 				...helpers,
 				getCookie,
@@ -38,14 +38,12 @@ export default function createHttpClient(settings, getStore, options = {}) {
 	}
 
 	return new HttpClient({
-		on_before_send,
-		catch_to_retry,
-		get_access_token,
-		transform_url               : settings.http.transformURL,
-		// `allowAbsoluteURLs` flag is deprecated.
-		allow_absolute_urls         : settings.http.allowAbsoluteURLs,
-		parseDates                  : settings.parseDates,
-		authentication_token_header : settings.authentication.header,
+		onBeforeSend,
+		catchToRetry,
+		getAuthToken,
+		transformUrl: settings.http.transformURL,
+		parseDates: settings.parseDates,
+		authTokenHeader: settings.authentication.header,
 		...options
 	})
 }